requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
statsmodels>=0.14.0
scipy>=1.10.0
openai-agents
